Added validation to UpdateStubCommandHandler and added unit tests

+++ src/HttPlaceholder.Application/Stubs/Commands/UpdateStubCommand/UpdateStubCommandHandler.cs
@@ -1,5 +1,7 @@
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using HttPlaceholder.Application.Exceptions;
 using HttPlaceholder.Application.StubExecution;
 using MediatR;
 
@@ -8,14 +10,22 @@ namespace HttPlaceholder.Application.Stubs.Commands.UpdateStubCommand
     public class UpdateStubCommandHandler : IRequestHandler<UpdateStubCommand>
     {
         private readonly IStubContext _stubContext;
+        private readonly IStubModelValidator _stubModelValidator;
 
-        public UpdateStubCommandHandler(IStubContext stubContext)
+        public UpdateStubCommandHandler(IStubContext stubContext, IStubModelValidator stubModelValidator)
         {
             _stubContext = stubContext;
+            _stubModelValidator = stubModelValidator;
         }
 
         public async Task<Unit> Handle(UpdateStubCommand request, CancellationToken cancellationToken)
         {
+            var validationResults = _stubModelValidator.ValidateStubModel(request.Stub);
+            if (validationResults.Any())
+            {
+                throw new ValidationException(validationResults);
+            }
+
             // Delete stub with same ID.
             await _stubContext.DeleteStubAsync(request.StubId);
             await _stubContext.DeleteStubAsync(request.Stub.Id);
